backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
pymongo==4.5.0
motor==3.3.2

# ML and Deep Learning
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
clip @ git+https://github.com/openai/CLIP.git
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Image processing
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
ujson==5.8.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1